refactor: Move ProxyFactoryConfigV2 into v3_proxyfactory package
