+++ proxy/src/main/java/hello/proxy/ProxyApplication.java
@@ -1,14 +1,5 @@
 package hello.proxy;
 
-import hello.proxy.config.AppV1Config;
-import hello.proxy.config.AppV2Config;
-import hello.proxy.config.v1_proxy.ConcreteProxyConfig;
-import hello.proxy.config.v1_proxy.InterfaceProxyConfig;
-import hello.proxy.config.v2_dynamicproxy.DynamicProxyBasicConfig;
-import hello.proxy.config.v2_dynamicproxy.DynamicProxyFilterConfig;
-import hello.proxy.config.v3_proxyfactory.ProxyFactoryConfigV1;
-import hello.proxy.config.v3_proxyfactory.advice.ProxyFactoryConfigV2;
-import hello.proxy.config.v4_postprocessor.BeanPostProcessorConfig;
 import hello.proxy.config.v5_autoproxy.AutoProxyConfig;
 import hello.proxy.trace.logtrace.LogTrace;
 import hello.proxy.trace.logtrace.ThreadLocalLogTrace;
